+++ src/imio/smartweb/core/contents/rest/events/endpoint.py
@@ -16,14 +16,19 @@ class BaseEventsEndpoint(BaseEndpoint):
         results = super(BaseEventsEndpoint, self).__call__()
         if not results.get("items"):
             return results
+        orientation = self.context.orientation
         for result in results["items"]:
             modified_hash = hash_md5(result["modified"])
             if result.get("image"):
-                self.convert_cached_image_scales(result, modified_hash)
+                self.convert_cached_image_scales(
+                    result, modified_hash, orientation=orientation
+                )
             for sub_content in result.get("items", []):
                 if sub_content["@type"] != "Image":
                     continue
-                self.convert_cached_image_scales(sub_content, modified_hash)
+                self.convert_cached_image_scales(
+                    sub_content, modified_hash, "image", ["preview"], ""
+                )
         return results
 
     @property