+++ src/imio/smartweb/core/webcomponents/src/components/Annuaire/ContactContent/ContactContent.jsx
@@ -327,7 +327,7 @@ const ContactContent = ({ queryUrl, onChange }) => {
                             <div className="spotlight-group flexbin r-content-gallery">
                                 {gallery.map((image) => (
                                     <a className="spotlight" href={image.image_full_scale} >
-                                        <img src={image.image_vignette_scale} />
+                                        <img src={image.image_preview_scale} />
                                     </a>
                                 ))}
                             </div>