+++ src/imio/smartweb/core/vocabularies.py
@@ -138,8 +138,8 @@ BootstrapCSSVocabulary = BootstrapCSSVocabularyFactory()
 class OrientationVocabularyFactory:
     def __call__(self, context=None):
         values = [
-            ("portrait", _("Portrait")),
             ("paysage", _("Paysage")),
+            ("portrait", _("Portrait")),
         ]
         terms = [
             SimpleVocabulary.createTerm(value[0], value[0], value[1])